requirements.txt
groq
watchdog
psycopg2-binary
